chore: fix js lib
