+++ test/app.ts
@@ -1,22 +1,26 @@
 import { createApp } from "../src/index.js";
 
-const app = createApp();
+let app = createApp();
 
 app.error(async (error, req, res) => {
     await Promise.resolve();
-    console.error("Error observed in error handler:", error, req, res);
+    console.error("Error", error, req, res);
 });
 
 app.get("/", (req, res) => {
+
     res.json({
         ok: true,
+        data: req.query,
     });
 });
 
 
 
-
 const server = await app.listen({
-    port: 8190
+    port: 8190,
+    
+    opt: { notify: true}
 });
+
 console.log(`http-native listening on ${server.url}`);
